fix: Rex 터레인 UV 정밀도 개선 - scale-bias 행렬 float32 → double 승격
문제: TMS 타일맵 깊은 줌 레벨에서 UV 매핑 정확도 저하.
원인: scale-bias 행렬이 CPU 전체 파이프라인에서 float32(osg::Matrixf)로
      저장되어 비-power-of-2 바이어스 값을 포함하는 경우(텍스처 아틀라스,
      비정렬 프로파일) float32 분해능 한계에 의한 정밀도 손실 발생.
      또한 addStandaloneImageLayer에서 TextureWindow 행렬과 누적
      scale-bias 행렬이 합성되지 않고 덮어쓰이는 버그 존재.

수정:
- TextureWindow::_matrix: osg::Matrixf → osg::Matrixd
- TerrainTileModel::Tile::matrix: osg::Matrixf → osg::Matrixd
- Sampler::_matrix: osg::Matrixf → osg::Matrixd
- Sampler::inheritFrom() 파라미터: Matrixf → Matrixd
- TileNode scaleBias[4]: osg::Matrixf[4] → osg::Matrixd[4]
- TerrainTileModelFactory addStandalone*: 내부 누적 행렬 Matrixd 사용
- addStandaloneImageLayer: TextureWindow 행렬과 parent-walking 행렬을
  postMult로 올바르게 합성 (기존 버그 수정)
- GPU 업로드 경계(COPY_MAT4F, glUniformMatrix4fv): Matrixd → float32
  명시적 변환 (GPU SSBO/uniform은 float32 유지)

+++ src/osgEarthDrivers/engine_rex/TileNode.cpp
@@ -26,12 +26,13 @@ using namespace osgEarth::Util;
 namespace
 {
     // Scale and bias matrices, one for each TileKey quadrant.
-    const osg::Matrixf scaleBias[4] =
+    // Stored as double precision to maintain UV accuracy at deep zoom levels.
+    const osg::Matrixd scaleBias[4] =
     {
-        osg::Matrixf(0.5f,0,0,0, 0,0.5f,0,0, 0,0,1.0f,0, 0.0f,0.5f,0,1.0f),
-        osg::Matrixf(0.5f,0,0,0, 0,0.5f,0,0, 0,0,1.0f,0, 0.5f,0.5f,0,1.0f),
-        osg::Matrixf(0.5f,0,0,0, 0,0.5f,0,0, 0,0,1.0f,0, 0.0f,0.0f,0,1.0f),
-        osg::Matrixf(0.5f,0,0,0, 0,0.5f,0,0, 0,0,1.0f,0, 0.5f,0.0f,0,1.0f)
+        osg::Matrixd(0.5,0,0,0, 0,0.5,0,0, 0,0,1.0,0, 0.0,0.5,0,1.0),
+        osg::Matrixd(0.5,0,0,0, 0,0.5,0,0, 0,0,1.0,0, 0.5,0.5,0,1.0),
+        osg::Matrixd(0.5,0,0,0, 0,0.5,0,0, 0,0,1.0,0, 0.0,0.0,0,1.0),
+        osg::Matrixd(0.5,0,0,0, 0,0.5,0,0, 0,0,1.0,0, 0.5,0.0,0,1.0)
     };
 }
 
@@ -1188,7 +1189,7 @@ TileNode::refreshInheritedData(TileNode* parent, const RenderBindings& bindings)
             {
                 Sampler& mySampler = myPass->sampler(SamplerBinding::COLOR_PARENT);
                 const Sampler& parentColorSampler = parentPass.sampler(SamplerBinding::COLOR);
-                osg::Matrixf newMatrix = parentColorSampler._matrix;
+                osg::Matrixd newMatrix = parentColorSampler._matrix;
                 newMatrix.preMult(scaleBias[quadrant]);
 
                 // Did something change?
